# core/types: add handcrafted and generated rlp encoders