@fjl 2025-03-14
One thing I don't like about this, is that we are losing the error messages with this change.