@s1na 2025-05-20
@MariusVanDerWijden Can you submit the StateAccount decoder in a separate PR? The decoder is quite simple there it has decent gains and the errors can be added in.