@MariusVanDerWijden 2025-05-21
@s1na I realized that I had a bug in the state account encoder, can't replicate any decent speedup anymore. Closing this for now